code box 예시 수정
